requirements.txt
asgiref==3.9.1
Django==5.2.4
sqlparse==0.5.3
djangorestframework==3.15.2
django-cors-headers==4.3.1
python-decouple==3.8
gunicorn==21.2.0
